Refactor adapters and patches once again
