+++ mlflow_cratedb/patch/mlflow/db_utils.py
@@ -2,14 +2,12 @@ import functools
 
 import sqlalchemy as sa
 
-from mlflow_cratedb.adapter.db import enable_refresh_after_dml
 from mlflow_cratedb.patch.sqlalchemy import patch_sqlalchemy_inspector
 
 
 def patch_db_utils():
     import mlflow.store.db.utils as db_utils
 
-    enable_refresh_after_dml()
     db_utils._initialize_tables = _initialize_tables
     db_utils._verify_schema = _verify_schema
 
@@ -22,7 +20,7 @@ def _initialize_tables(engine: sa.Engine):
     """
     from mlflow.store.db.utils import _logger
 
-    from mlflow_cratedb.adapter.db import _setup_db_create_tables
+    from mlflow_cratedb.adapter.setup_db import _setup_db_create_tables
 
     patch_sqlalchemy_inspector(engine)
     _logger.info("Creating initial MLflow database tables...")
